requirements.txt
# ConanBot — Python requirements
# Notes:
# - Keep this file as the single source of truth for Python deps.
# - Some features are optional; comment out lines you don't need yet.
# - System-level dependencies (Tesseract, FFmpeg, etc.) are listed in EXTERNAL_DEPENDENCIES.md.

# === CORE (SDK & events) ===
pydantic>=2.7.0
ulid-py>=1.1.0

# === DATA & NUMERIC ===
numpy>=1.26
pandas>=2.2

# === VISION / ML (comment out if not training yet) ===
# Select a Torch build suitable for your platform/CUDA.
# See https://pytorch.org/get-started/locally/ for the exact command if you need CUDA.
torch>=2.3
torchvision>=0.18
pillow>=10.3.0
opencv-python>=4.10

# === AUDIO CAPTURE ===
sounddevice>=0.4
pyaudio>=0.2.14

# === OCR ===
pytesseract>=0.3.10

# === CONFIG / YAML ===
pyyaml>=6.0.1

# === SYSTEM & SCREEN CAPTURE (Windows) ===
psutil>=5.9
mss>=9.0.1
dxcam>=0.0.5 ; platform_system == "Windows"
pywin32>=306 ; platform_system == "Windows"

# === INPUTS (controllers/keyboard) ===
inputs>=0.5
pynput>=1.7.7

# === VIDEO ENCODING / FFMPEG (optional) ===
av>=12.3
ffmpeg-python>=0.2.0

# === UI (FastAPI) ===
fastapi>=0.115
uvicorn[standard]>=0.30
websockets>=12.0
sse-starlette>=2.1.3

# === STREAMING SINKS (Kafka / S3) ===
# Choose ONE Kafka client (confluent-kafka preferred):
confluent-kafka>=2.4.0
# kafka-python>=2.0.2
boto3>=1.34
aioboto3>=12.4.0
s3fs>=2024.6.0
smart_open>=7.0.4

# === RL / PPO ===
stable-baselines3>=2.3.0
gymnasium>=0.29
tensorboard>=2.17.0

# === DEV / DX ===
rich>=13.7
typer>=0.12.5
pytest>=8.2

# === OPTIONAL PERFORMANCE EXTRAS ===
# Faster JSON (used by FastAPI for responses if configured)
# orjson>=3.10.0
# Audio utilities (if doing signal processing)
# soundfile>=0.12.1
# scipy>=1.13
# Async test helpers for FastAPI/WebSockets
# pytest-asyncio>=0.23
